Polishing and package renaming to wso2
