--- components/analytics-data-connectors/es.eci.carbon.analytics.datasource.mongo/src/main/java/es/eci/carbon/analytics/datasource/mongo/MongoConstants.java
@@ -1,11 +1,0 @@
-package es.eci.carbon.analytics.datasource.mongo;
-
-public class MongoConstants {
-
-	public static final String DATASOURCE_NAME = "datasource";
-
-	public static final String DATABASE = "database";
-
-	public static final String DEFAULT_ARS_DB_NAME = "ARS";
-
-}
