--- components/analytics-data-connectors/es.eci.carbon.analytics.datasource.mongo/src/test/resources/conf1/analytics/analytics-config.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<analytics-dataservice-configuration>
-    <primaryRecordStore>EVENT_STORE</primaryRecordStore>
-    <!-- Analytics Record Store - properties related to record storage implementation -->
-    <analytics-record-store name="EVENT_STORE">
-        <implementation>es.eci.carbon.analytics.datasource.mongo.MongoAnalyticsRecordStore</implementation>
-        <properties>
-            <property name="datasource">WSO2_ANALYTICS_DS_MONGO</property>
-            <property name="database">ARS_TEST</property>
-        </properties>
-    </analytics-record-store>
-    <!-- The data indexing analyzer implementation -->
-    <analytics-lucene-analyzer>
-        <implementation>org.apache.lucene.analysis.standard.StandardAnalyzer</implementation>
-    </analytics-lucene-analyzer>
-    <!-- The number of index shards, should be equal or higher to the number of indexing nodes that is going to be working,
-         ideal count being 'number of indexing nodes * CPU cores' -->
-    <shardCount>5</shardCount>
-</analytics-dataservice-configuration>
