+++ components/analytics-data-connectors/org.wso2.carbon.analytics.datasource.mongo/src/main/java/org/wso2/carbon/analytics/datasource/mongo/MongoConstants.java
@@ -1,0 +1,11 @@
+package org.wso2.carbon.analytics.datasource.mongo;
+
+public class MongoConstants {
+
+	public static final String DATASOURCE_NAME = "datasource";
+
+	public static final String DATABASE = "database";
+
+	public static final String DEFAULT_ARS_DB_NAME = "ARS";
+
+}
