+++ components/analytics-data-connectors/pom.xml
@@ -35,6 +35,7 @@
         <module>org.wso2.carbon.analytics.datasource.rdbms</module>
         <module>org.wso2.carbon.analytics.datasource.hbase</module>
         <module>org.wso2.carbon.analytics.datasource.cassandra</module>
+	<module>org.wso2.carbon.analytics.datasource.mongo</module>
     </modules>
 
 </project>
